requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
psycopg2-binary==2.9.2
requests==2.26.0
PyYAML==6.0
uritemplate==4.1.1
coreapi==2.3.3
python-decouple==3.5